requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
playwright==1.36.0
pydantic==1.10.7
python-dotenv==1.0.0
pymongo==4.3.3
redis==4.5.4
loguru==0.7.0
jinja2==3.1.2
openai==0.27.8
celery==5.2.7
httpx==0.24.0
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-playwright==0.3.3 